Add picom compositor
This seems to fix the weird flickering behavior I was seeing, where
the whole display would, for the duration of a moment, "snap back" to
an earlier state. It was hard to reproduce, (I mostly saw it while
typing in a terminal or textbox in firefox) but using godot_4 and then
changing virtual desktops rapidly in i3 more reliably elicited it.
I assume this was due to heavier gpu use.

The popup context menus in firefox now look different. They have a
black border like alpha-blending isn't working right, but I can live
with that.